Refactor: Page 구조 변경
