+++ src/components/chatbot/pages/ChatbotPage/ChatbotPage.tsx
@@ -1,0 +1,171 @@
+'use client';
+
+import { Send } from 'lucide-react';
+import { v4 as uuidv4 } from 'uuid';
+
+import { ChangeEvent, FormEvent, useEffect, useRef, useState } from 'react';
+
+import logo from '@/assets/logo.png';
+import { Chat } from '@/components/chatbot/organisms/ChatItem';
+import { HeadlineBanner, HeadlineBannerSkeleton } from '@/components/chatbot/organisms/HeadlineBanner';
+import { Button } from '@/components/common/atoms/Button';
+import { SafeImage } from '@/components/common/atoms/SafeImage';
+import { Input } from '@/components/common/molecules/Input';
+import { useHeadlineData } from '@/hooks/useGetHeadLine';
+import { EventName } from '@/lib/firebase/eventNames';
+import { firebaseLogging } from '@/lib/firebase/logEvent';
+import { usePostChatbotQuestion } from '@/queries/chatbot/usePostChatbotQuestion';
+import { useToastStore } from '@/stores/toastStore';
+import { ChatMessage } from '@/types/chatbot/chatMessage';
+
+const suggestQuestions: string[] = ['휴가 신청하는 법을 알려줘.', '유료 구독료 지원 일정을 알려줘.'];
+
+export function ChatbotPage() {
+  const [chats, setChats] = useState<ChatMessage[]>([]);
+  const [currentInput, setCurrentInput] = useState('');
+  // 상단 헤드라인 배너 가져오기
+  const { headlines, isLoading: isHeadlineLoading } = useHeadlineData();
+
+  const { mutateAsync: postQuestion, isPending } = usePostChatbotQuestion();
+
+  const { showToast } = useToastStore();
+
+  const messagesEndRef = useRef<HTMLDivElement>(null);
+
+  // 추천 검색어 클릭
+  const handleSuggestQuestionClick = (text: string) => {
+    firebaseLogging(EventName.CHATBOT_RECOMMEND_QUESTION_CLICK, {
+      description: text,
+    });
+    loadAnswer(text);
+  };
+
+  // 채팅 입력
+  function handleInputChange(e: ChangeEvent<HTMLInputElement>) {
+    const newValue = e.target.value;
+    if (newValue.length > 500) {
+      showToast('메시지는 최대 500자까지 입력 가능합니다.', 'error', 'top');
+      return;
+    }
+    setCurrentInput(newValue);
+  }
+
+  // 채팅 전송 버튼 클릭
+  const handleSendClick = (e: FormEvent) => {
+    e.preventDefault();
+    firebaseLogging(EventName.CHATBOT_QUESTION_SEND_CLICK);
+    loadAnswer(currentInput);
+  };
+
+  // 현재 채팅을 chat 리스트에 등록하고, 답변 받아옴
+  const loadAnswer = async (text: string) => {
+    setChats((prev) => [
+      ...prev,
+      {
+        id: uuidv4(),
+        text: text,
+        type: 'question',
+      },
+    ]);
+
+    setCurrentInput('');
+    const res = await postQuestion(text);
+    if (res && res.data) {
+      setChats((prev) => [
+        ...prev,
+        {
+          id: uuidv4(),
+          text: res.data.answer,
+          type: 'answer',
+        },
+      ]);
+    }
+  };
+
+  useEffect(() => {
+    setTimeout(() => {
+      // chat 갱신된다면, 제일 밑으로 스크롤
+      messagesEndRef.current?.scrollIntoView({ behavior: 'smooth' });
+    }, 0);
+  }, [chats]);
+
+  return (
+    <div className="flex flex-col min-h-app bg-neutral-50">
+      {/* 최신 공지와 뉴스 */}
+      {isHeadlineLoading && chats.length === 0 && <HeadlineBannerSkeleton />}
+      {!isHeadlineLoading && chats.length === 0 && headlines.length > 0 && <HeadlineBanner items={headlines} />}
+
+      <div className="flex-1 flex flex-col gap-4 p-4 pb-16 relative">
+        {/* 로고와 추천 질문 */}
+        {chats.length === 0 && (
+          <div className="flex-1 flex flex-col items-center justify-center gap-4">
+            <div className="mb-2 transform hover:scale-105 transition-transform duration-300 px-3">
+              <SafeImage src={logo} alt="logo" width={400} height={50} priority className="drop-shadow-md" />
+            </div>
+
+            <div className="w-full max-w-md space-y-3">
+              {suggestQuestions.map((text) => (
+                <div
+                  key={text}
+                  onClick={() => handleSuggestQuestionClick(text)}
+                  className="bg-white border border-gray-200 p-3 rounded-2xl text-sm text-gray-700 hover:bg-gray-100 hover:border-gray-300 cursor-pointer shadow-sm transition-all duration-200"
+                >
+                  {text}
+                </div>
+              ))}
+            </div>
+          </div>
+        )}
+
+        {/* 채팅 메시지 */}
+        <div className="space-y-4">
+          <Chat>
+            {chats.map((chat) =>
+              chat.type === 'question' ? (
+                <Chat.UserMessage key={chat.id} text={chat.text} />
+              ) : (
+                <Chat.BotMessage key={chat.id} text={chat.text} />
+              ),
+            )}
+
+            {isPending && <Chat.BotLoading />}
+          </Chat>
+        </div>
+
+        <div ref={messagesEndRef} />
+      </div>
+
+      {/* 입력 영역 */}
+      <form className="relative bottom-0 w-full max-w-app">
+        <div className="bg-white p-4 border-t border-gray-100  rounded-t-3xl">
+          <div className="flex items-center bg-white rounded-full pr-2 gap-2 shadow-md">
+            <Input
+              type="text"
+              placeholder="질문을 입력하세요"
+              value={currentInput}
+              onChange={handleInputChange}
+              className="flex-1 text-gray-700 rounded-full"
+              autoFocus={true}
+            />
+
+            <Button
+              type="submit"
+              size="icon"
+              className={`text-white shrink-0 ${
+                isPending || currentInput === '' ? 'opacity-50 cursor-not-allowed' : ''
+              }`}
+              onClick={handleSendClick}
+              disabled={isPending || currentInput === ''}
+            >
+              <Send size={18} />
+            </Button>
+          </div>
+
+          {/*<p className="text-xs text-center text-gray-400 mt-2">*/}
+          {/*  2025년 4월 7까지 업데이트된 챗봇입니다.*/}
+          {/*</p>*/}
+        </div>
+      </form>
+    </div>
+  );
+}
