+++ src/components/chatbot/pages/ChatbotPage/index.ts
@@ -1,0 +1,1 @@
+export * from './ChatbotPage';
