+++ src/components/post/pages/NewsListPage/index.tsx
@@ -1,0 +1,1 @@
+export * from './NewsListPage';
