+++ src/components/post/pages/NoticeDetailPage/index.ts
@@ -1,0 +1,1 @@
+export * from './NoticeDetailPage';
