+++ src/components/post/pages/NoticeListPage/index.ts
@@ -1,0 +1,1 @@
+export * from './NoticeListPage';
